move connectivity to experimental
