+++ lib/iris/tests/unit/common/metadata/test_metadata_manager_factory.py
@@ -21,11 +21,11 @@ from iris.common.metadata import (
     AncillaryVariableMetadata,
     BaseMetadata,
     CellMeasureMetadata,
-    ConnectivityMetadata,
     CoordMetadata,
     CubeMetadata,
     metadata_manager_factory,
 )
+from iris.experimental.ugrid import ConnectivityMetadata
 
 
 BASES = [
