+++ lib/iris/tests/unit/experimental/ugrid/__init__.py
@@ -3,7 +3,4 @@
 # This file is part of Iris and is released under the LGPL license.
 # See COPYING and COPYING.LESSER in the root of the repository for full
 # licensing details.
-"""
-Ugrid code is tested in this package.
-
-"""
+"""Unit tests for the :mod:`iris.experimental.ugrid` package."""
