+++ requirements/all.txt
@@ -3,5 +3,4 @@
 mo_pack
 nc-time-axis
 pandas
-pyugrid
 stratify
